refactor: account deletion service
