+++ src/main/java/fast/bank/api/domain/account/service/validation/deletion/AccountDeletionValidators.java
@@ -1,4 +1,4 @@
-package fast.bank.api.domain.account.service.deletion.validation;
+package fast.bank.api.domain.account.service.validation.deletion;
 
 import fast.bank.api.domain.account.model.Account;
